perf: barrel export components
