+++ src/nimbus/components/cf/CompatibilityFlags.astro
@@ -5,7 +5,7 @@ import { getCollection, render } from "astro:content";
 import AnchorHeading from "./AnchorHeading.astro";
 // Forward the MDX globals so admonitions (`:::note` → `<Aside>`) resolve inside
 // each rendered flag entry, the same way Render.astro / ChangelogEntry.astro do.
-import { components } from "@/components";
+import { components } from "@/mdx-components";
 
 type Props = z.infer<typeof props>;
