+++ src/nimbus/components/changelog/ChangelogEntry.astro
@@ -11,7 +11,7 @@
 import { render, type CollectionEntry } from "astro:content";
 import DateRail from "./DateRail.astro";
 import ProductPills from "./ProductPills.astro";
-import { components } from "@/components";
+import { components } from "@/mdx-components";
 
 interface Props {
   entry: CollectionEntry<"changelog">;
