+++ src/nimbus/components/index.ts
@@ -1,3 +1,1 @@
-// Re-export the MDX component barrel so the trailing-slash form `~/components/`
-// (used by a few content files) resolves the same as `~/components`.
 export * from "../components";
