+++ src/nimbus/mdx-components.ts
@@ -1,0 +1,103 @@
+import { Aside } from "./components/ui/aside";
+import { Badge } from "./components/ui/badge";
+import { Card } from "./components/ui/card";
+import { CardGrid } from "./components/ui/card-grid";
+import { Code } from "./components/ui/code";
+import { FileTree } from "./components/ui/file-tree";
+import { PackageManagers } from "./components/ui/package-managers";
+import { Step, Steps } from "./components/ui/steps";
+import { TabItem, Tabs } from "./components/ui/tabs";
+import LinkButton from "./components/ui/link-button/LinkButton.astro";
+import Render from "./components/Render.astro";
+import APIRequest from "./components/cf/APIRequest.astro";
+import DashButton from "./components/cf/DashButton.astro";
+import DirectoryListing from "./components/cf/DirectoryListing.astro";
+import Description from "./components/cf/Description.astro";
+import Details from "./components/cf/Details.astro";
+import MetaInfo from "./components/cf/MetaInfo.astro";
+import Type from "./components/cf/Type.astro";
+import WranglerConfig from "./components/cf/WranglerConfig.astro";
+import WranglerNamespace from "./components/cf/WranglerNamespace.astro";
+import Feature from "./components/cf/Feature.astro";
+import Glossary from "./components/cf/Glossary.astro";
+import GlossaryTooltip from "./components/cf/GlossaryTooltip.astro";
+import LinkCard from "./components/cf/LinkCard.astro";
+import LinkTitleCard from "./components/cf/LinkTitleCard.astro";
+import ListTutorials from "./components/cf/ListTutorials.astro";
+import Plan from "./components/cf/Plan.astro";
+import ProductReleaseNotes from "./components/cf/ProductReleaseNotes.astro";
+import ProductChangelog from "./components/cf/ProductChangelog.astro";
+import RelatedProduct from "./components/cf/RelatedProduct.astro";
+import TypeScriptExample from "./components/cf/TypeScriptExample.astro";
+import InlineBadge from "./components/cf/InlineBadge.astro";
+import YouTube from "./components/cf/YouTube.astro";
+import Example from "./components/cf/Example.astro";
+import Markdown from "./components/cf/Markdown.astro";
+import CURL from "./components/cf/CURL.astro";
+import GitHubCode from "./components/cf/GitHubCode.astro";
+import Width from "./components/cf/Width.astro";
+import RuleID from "./components/cf/RuleID.astro";
+import PublicStats from "./components/cf/PublicStats.astro";
+import RSSButton from "./components/cf/RSSButton.astro";
+import GlossaryDefinition from "./components/cf/GlossaryDefinition.astro";
+import WranglerCommand from "./components/cf/WranglerCommand.astro";
+import AnchorHeading from "./components/cf/AnchorHeading.astro";
+import FeatureTable from "./components/cf/FeatureTable.astro";
+import PagesBuildPreset from "./components/cf/PagesBuildPreset.astro";
+import AvailableNotifications from "./components/cf/AvailableNotifications.astro";
+import Stream from "./components/cf/Stream.astro";
+import ResourcesBySelector from "./components/cf/ResourcesBySelector.astro";
+
+export const components = {
+  APIRequest,
+  AnchorHeading,
+  Aside,
+	AvailableNotifications,
+  Badge,
+  CURL,
+  Card,
+  CardGrid,
+  Code,
+  DashButton,
+  Description,
+  Details,
+  DirectoryListing,
+  Example,
+  Feature,
+  FeatureTable,
+  FileTree,
+  GitHubCode,
+  Glossary,
+  GlossaryDefinition,
+  GlossaryTooltip,
+  InlineBadge,
+  LinkButton,
+  LinkCard,
+  LinkTitleCard,
+  ListTutorials,
+  Markdown,
+  MetaInfo,
+  PackageManagers,
+  PagesBuildPreset,
+  Plan,
+  ProductChangelog,
+  ProductReleaseNotes,
+  PublicStats,
+  RelatedProduct,
+  Render,
+  ResourcesBySelector,
+  RSSButton,
+  RuleID,
+  Step,
+  Steps,
+  Stream,
+  TabItem,
+  Tabs,
+  Type,
+  TypeScriptExample,
+  Width,
+  WranglerCommand,
+  WranglerConfig,
+  WranglerNamespace,
+  YouTube,
+};
