+++ src/nimbus/pages/[...slug].astro
@@ -12,7 +12,7 @@ import {
 } from "nimbus-docs";
 import { config } from "virtual:nimbus/config";
 import { docsSidebarTransform } from "../util/sidebar";
-import { components } from "../components";
+import { components } from "../mdx-components";
 
 export const prerender = true;
 export const getStaticPaths = getDocsStaticPaths;
