+++ src/nimbus/pages/changelog/post/[...slug].astro
@@ -11,7 +11,7 @@ import { render } from "astro:content";
 
 import ChangelogLayout from "~/layouts/ChangelogLayout.astro";
 import ProductPills from "~/components/changelog/ProductPills.astro";
-import { components } from "~/components";
+import { components } from "~/mdx-components";
 
 export const prerender = true;
