+++ src/nimbus/pages/waf/change-log/changelog/[...page].astro
@@ -14,7 +14,7 @@ import { externalAppLinksTransform } from "~/util/sidebar";
 import Pagination from "~/components/changelog/Pagination.astro";
 import RSSButton from "~/components/cf/RSSButton.astro";
 import AnchorHeading from "~/components/cf/AnchorHeading.astro";
-import { components } from "@/components";
+import { components } from "@/mdx-components";
 import { getChangelogs } from "~/util/changelog";
 
 export const prerender = true;
